optimize output of running dodcr without command
